# MyProject/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for retrieving cryptocurrency exchange rates and statistics on average, minimum, and maximum values.",
        "title": "MyProject Crypto API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/coins/get_actual/{titles}": {
            "get": {
                "description": "Returns the latest actual rates for the specified currencies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "coins"
                ],
                "summary": "Fetch actual currency rates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "List of currencies separated by commas (e.g., BTC,ETH,USD)",
                        "name": "titles",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with actual rates",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CoinDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request (missing titles parameter)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No coins found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CoinDTO": {
            "type": "object",
            "properties": {
                "rate": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}